+++ passes/S007/testdata/src/a/alias.go
@@ -1,0 +1,19 @@
+package a
+
+import (
+	s "github.com/hashicorp/terraform/helper/schema"
+)
+
+func falias() {
+	_ = s.Schema{ // want "schema should not enable Required and configure ConflictsWith"
+		Required:      true,
+		ConflictsWith: []string{},
+	}
+
+	_ = map[string]*s.Schema{
+		"name": { // want "schema should not enable Required and configure ConflictsWith"
+			Required:      true,
+			ConflictsWith: []string{},
+		},
+	}
+}