+++ passes/S007/testdata/src/a/comment_ignore.go
@@ -1,0 +1,13 @@
+package a
+
+import (
+	"github.com/hashicorp/terraform/helper/schema"
+)
+
+func fcommentignore() {
+	//lintignore:S007
+	_ = schema.Schema{
+		Required:      true,
+		ConflictsWith: []string{},
+	}
+}