+++ passes/S007/testdata/src/a/outside_package.go
@@ -1,0 +1,19 @@
+package a
+
+import (
+	"a/schema"
+)
+
+func foutside() {
+	_ = schema.Schema{
+		Required:      true,
+		ConflictsWith: []string{},
+	}
+
+	_ = map[string]*schema.Schema{
+		"name": {
+			Required:      true,
+			ConflictsWith: []string{},
+		},
+	}
+}