+++ passes/S007/testdata/src/a/schema/schema.go
@@ -1,0 +1,6 @@
+package schema
+
+type Schema struct {
+	ConflictsWith []string
+	Required      bool
+}